+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Releases
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Export a mapping of constants in the generated module under the ``constants``
+  attribute.
 
 
 0.1.0 (2015-08-28)